backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
numpy==1.26.3
scikit-learn==1.4.0
nltk==3.8.1
stix2==3.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
ollama
xgboost==2.0.3
pandas==2.2.0
scapy==2.7.0
Jinja2==3.1.6
xhtml2pdf==0.2.17
